Added raycast
