--- Journey/Assets/Prefab/Bullet.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0ea584959d1a2084dac9d796ee3850d2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
