--- Journey/Assets/Scripts/Player_move.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 49e5ec5161f70174c85aa35e55ee5992
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
